distributions/hyper: support for the hypergeometric distribution
The hypergeometric distribution is a discrete probability distribution
that describes the probability of k successes (random draws for which
the object drawn has a specified feature) in n draws, without
replacement, from a finite population of size N that contains exactly
K objects with that feature, wherein each draw is either a success or
a failure.

Added documentation, examples and tests for the hypergeometric
distribution.

All features implemented according to https://en.wikipedia.org/wiki/Hypergeometric_distribution